Extract "pwm_handler" into separate module.
